ML: rateLimiter.c - change the type of the drain_rate: from 'int' to 'int64_t'. Solves the problem of overflow on 32-bits systems.
git-svn-id: https://repository.napa-wine.eu/svn/napawine-software/trunk/NAPA-BASELIBS@2363 c54fac2e-1815-407e-9b96-088d06743913

+++ ml/util/rateLimiter.c
@@ -21,7 +21,8 @@
 
 extern struct event_base *base;
 static long bucket_size = 0;
-static int drain_rate = 0;
+static int64_t drain_rate = 0;
+
 
 static long bytes_in_bucket = 0;
 struct timeval bib_then = { 0, 0};
